video_digest/backend/requirements.txt
Flask
python-dotenv
youtube-transcript-api
requests
pydub
google-generativeai
